withBlock() moved to CommonApiFunctions, files split & renaming in scorex-transaction api/http
