+++ scorex-basics/src/main/scala/scorex/api/http/CommonApiFunctions.scala
@@ -1,7 +1,10 @@
 package scorex.api.http
 
-import play.api.libs.json.Json
+import play.api.libs.json.{JsValue, Json}
 import akka.util.Timeout
+import scorex.block.Block
+import scorex.crypto.Base58
+import scorex.transaction.History
 import scala.concurrent.duration._
 
 
@@ -9,4 +12,14 @@ trait CommonApiFunctions {
   implicit val timeout = Timeout(5.seconds)
 
   def json(t: Throwable) = Json.obj("error" -> Unknown.id, "message" -> t.getMessage)
+
+  protected[api] def withBlock(encodedSignature: String)
+                              (action: Block => JsValue)
+                              (implicit history: History): JsValue =
+    Base58.decode(encodedSignature).toOption.map { signature =>
+      history.blockById(signature) match {
+        case Some(block) => action(block)
+        case None => BlockNotExists.json
+      }
+    }.getOrElse(InvalidSignature.json)
 }
